+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Added:
 * CNVpytor plots into the CNV PDF report https://github.com/Clinical-Genomics/BALSAMIC/pull/1023
 * Research and clinical housekeeper tags https://github.com/Clinical-Genomics/BALSAMIC/pull/1023
 * Cluster configuration for rules https://github.com/Clinical-Genomics/BALSAMIC/pull/1028
+* Variant filteration using loqusDB and Swegen annotations https://github.com/Clinical-Genomics/BALSAMIC/pull/1029
 
 Changed:
 ^^^^^^^^